pyproject.toml
[tool.poetry]
name = "bscproautomationbot"
version = "0.1.0"
description = ""
authors = ["Roderick Buo <roderickbuo31@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
gspread = "^4.0.1"
selenium = "^3.141.0"
google-api-python-client = "^2.23.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"
webdriver-manager = "^3.4.2"
pandas = "1.3.3"
oauth2client = "^4.1.3"
schedule = "^1.1.0"
numpy = "1.17.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
